pyproject.toml
[project]
name = "adventofcode2025"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
